Backend/sensors_to_mqtt/requirements.txt
apache-beam
paho-mqtt
requests